2008-02-27  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Fixed compile error on Mac OS X, on 64Bit OS maybe.
	* src/DHTBucket.cc

	Removed unnecessary cast.
	* src/HandshakeExtensionMessage.cc

+++ src/DHTBucket.cc
@@ -46,7 +46,7 @@
 
 namespace aria2 {
 
-DHTBucket::DHTBucket(uint32_t prefixLength,
+DHTBucket::DHTBucket(size_t prefixLength,
 		     const unsigned char* max, const unsigned char* min,
 		     const SharedHandle<DHTNode>& localNode):
   _prefixLength(prefixLength),
